Simplify graphcms fetching
