+++ src/modules/core/gql/types/StaticDataset.ts
@@ -1,0 +1,6 @@
+import { Customer } from '@/modules/core/data/types/Customer';
+
+export type StaticCustomer = Pick<Customer, 'availableLanguages'>;
+export type StaticDataset = {
+  customer: StaticCustomer;
+};
